Update main.js
Using npm start: encountered the "TypeError: this.createWindow is not a function" due to the context of this when calling this.createWindow() inside your constructor init() method.

Context Binding: Ensure that you use arrow functions or explicitly bind the this context. Since you're using a class, the best approach would be to convert any method that needs to retain this context into an arrow function.

Create Window Function: You need to explicitly define the createWindow function inside your class. Here is a modified version of your code: